Fix Ebenezer crash on ZONE_INFO.ServerNo (tinyint) cast
ServerNo is tinyint in the imported schema, but the reader called
GetInt16 on it. SqlDataReader requires an exact type match, so this
threw InvalidCastException on every startup, and
HostOptions.BackgroundServiceExceptionBehavior = StopHost took down
the whole Ebenezer process. AiServerDb already reads the same column
correctly with GetByte.

+++ dotnet/src/Servers/OpenKO.Servers.Ebenezer/Db/EbenezerDb.cs
@@ -45,7 +45,7 @@ public sealed class EbenezerDb(SqlConnectionFactory connectionFactory, ILogger l
             "ServerNo, ZoneNo, strZoneName, InitX, InitZ, InitY, Type, RoomEvent",
             static reader => new ZoneInfo
             {
-                ServerId = (byte)reader.GetInt16(0),
+                ServerId = reader.GetByte(0),
                 ZoneId = reader.GetInt16(1),
                 Name = reader.IsDBNull(2) ? string.Empty : reader.GetString(2).TrimEnd(),
                 InitX = reader.GetInt32(3),
